requirements.txt
onvif-zeep
opencv-python
pillow
appdirs